fix(ci): coverage shard — project required fields and split oversized tests
